Add first version of ProjConnect's documentation (#1)
* Add UML models and build script

The current version should be reviewed before implementation.

* Add way of getting UserData and Post and loggin in

Adding the way how the client side gets the UserData and also a Post
with a ID. Also the way which the login in request should behave.

* Improve folder's nameming and improve build script

Improving folder's naming from image and src-plantuml to img and src,
respectivily. Also build script was improved by making it more verbal
(it now tells if building a image was done succefully or not) and by
making it less hard coded with a for loop.

* Add documentation to backend and to database

Adding small description of the data models and some of the requests
should be handled by the server.

* Add main endpoints in server-client model

Adding main endpoints that will be available in server with exception of
the endpoint related with login and POST requests, that still need to be
discussed.